+++ src/form/form/index.js
@@ -1,1 +1,1 @@
-import './form'
+export { default } from './form'